requirements.txt
requests >= 2.31.0
azure-identity >= 1.15.0